refactoring remember_me.py and created a function called greet_user( )

+++ remember_me.py
@@ -1,11 +1,17 @@
 from pathlib import Path
 import json
 
-username = input("\nWhat's is your name? ")
-
+def greet_user():
+    """Greet the user by name.""" 
 path = Path('username.json')
-contents = json.dumps(username)
-path.write_text(contents)
+if path.exists():
+    contents = path.read_text()
+    username = json.loads(contents)
+    print(f"Welcome back, {username}!")
+else:
+    username = input("\nWhat's is your name? ")
+    contents = json.dumps(username)
+    path.write_text(contents)
+    print(f"\nWe'll remember you when you come back, {username}!")
 
-print(f"\nWe'll remember you when you come back, {username}!")
-
+greet_user()
